// cosmowcl/src/app/shared/models/api/cosmoauthorization/applicazioneEsternaConValidita.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * applicazioni-esterne.v1
 * CRUD per le applicazioni esterne e le funzionalita\' associate
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CampiTecnici } from './campiTecnici.d';
import { FunzionalitaApplicazioneEsterna } from './funzionalitaApplicazioneEsterna.d';


export interface ApplicazioneEsternaConValidita { 
    id?: number;
    descrizione: string;
    icona: string;
    funzionalitaPrincipale?: FunzionalitaApplicazioneEsterna;
    campiTecnici?: CampiTecnici;
    associataUtenti?: boolean;
    associataEnti?: boolean;
    numEntiAssociati?: number;
}